# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sweet Tooth — Pakistan's Home Baker Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers for a baker
      parameters:
        - name: bakerId
          in: query
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
  /customers/{customerId}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer details
      parameters:
        - name: customerId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  # ── ANALYTICS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Customer:
      type: object
      required: [id, name, whatsappNumber, totalOrders, totalSpentPkr, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        whatsappNumber: { type: string }
        city: { type: ["string", "null"] }
        preferredArea: { type: ["string", "null"] }
        totalOrders: { type: integer }
        totalSpentPkr: { type: integer }
        lastOrderAt: { type: ["string", "null"] }
        isRegular: { type: boolean }
        isAtRisk: { type: boolean }
        createdAt: { type: string }

    # ── Marketplace schemas ───────────────────────────────────────────────────
...